move diag helpers to own file; update diag rec to match field
