+++ odbc/src/api/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod definitions;
+pub(crate) mod diag;
 pub(crate) mod errors;
 mod functions;
 pub(crate) mod odbc_uri;
